requirements.txt
feedparser>=6.0
google-genai>=1.0.0
python-telegram-bot>=21.0
pyyaml>=6.0
